Refactor ballerina to openAPI code structure
