+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/mime_with_record_payload.yaml
