+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/multiple_payload.yaml
