+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/nested_array.yaml
