+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/rb_scenario02.yaml
