+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/rb_scenario06.yaml
