+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/record_field_array.yaml
