+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/arrayRecord_payload_service.bal
