+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/arrayRecordfield_payload_service.bal
