+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/json_payload_service.bal
