+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/mime_with_recordpayload_service.bal
