+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/multiple_payload_service.bal
