+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/nested2Record_payload_service.bal
