+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/nestedRecord_payload_service.bal
