+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/rb_scenario01.bal
