+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/record_payload_service.bal
