+++ openapi-cli/src/test/resources/ballerina-to-openapi/request_body/xml_payload_service.bal
