+++ openapi-cli/src/test/resources/testng.xml
@@ -26,6 +26,7 @@ under the License.
         <classes>
             <class name="io.ballerina.openapi.cmd.OpenAPICmdTest"/>
             <class name="io.ballerina.ballerina.OpenApiConverterUtilsTest"/>
+            <class name="io.ballerina.ballerina.RequestBodyTest"/>
             <class name="io.ballerina.openapi.cmd.OpenApiGenServiceCmdTest"/>
             <class name="io.ballerina.openapi.CodeGeneratorTest"/>
             <class name="io.ballerina.generators.common.GeneratorUtilsTests"/>
